requirements.txt
psycopg2-binary
fastapi[standard]
python-dotenv
pandas